+++ src/Ext/TechnoType/Body.cpp
@@ -374,6 +374,8 @@ void TechnoTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 		}
 	}
 
+	this->ImmuneToGarrisonPenetration.Read(exINI, pSection, "ImmuneToGarrisonPenetration");
+
 	// Art tags
 	INI_EX exArtINI(CCINIClass::INI_Art);
 	auto pArtSection = pThis->ImageFile;
@@ -875,6 +877,8 @@ void TechnoTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->Wake)
 		.Process(this->Wake_Grapple)
 		.Process(this->Wake_Sinking)
+
+		.Process(this->ImmuneToGarrisonPenetration)
 		.Process(this->Transporter_FixCyborgLegs)
 
 		.Process(this->NoManualUnload)