+++ docs/Whats-New.md
@@ -499,6 +499,7 @@ New:
 - Custom object palettes for TerrainTypes (by Starkku)
 - Forbidding parallel AI queues for specific TechnoTypes (by Starkku)
 - Nonprovocative Warheads (by Starkku)
+- Penetration damage on garrisonable structures (by FS-21)
 - Map action 507 for printing a message with the remaining map objects (by FS-21)
 - ScriptType actions `10018` & `12003` for aircrafts attacks (by FS-21)